requirements.txt
streamlit
transformers
torch
pandas
yfinance
scikit-learn
requests
python-dotenv